Add --use-first-bug option to bz attach disambiguate when multiple bugs are referenced in a commit message.
With the standard Mozilla style, the first one is almost certainly the one we want.

I did not make this the default because I think we hit this case if there are multiple attachments, each with their
own bug, and in that case I think it is better to fail.

I referenced the new option in the error message so that people are made aware of it when they need it.